refactor into class based
