+++ termgpt/models/__init__.py
@@ -1,0 +1,2 @@
+from termgpt.models.anthropic import CLAUDE, AnthropicChat
+from termgpt.models.openai import GPT3, GPT4, OpenAIChat
